+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v2.7.0
+- Auto update check
+- now shows setup button
+- fixed difficulty face position
+- queue count indicator
+- and more
+
 ## v2.6.0
 - added an option to copy youtube link to clipboard instead of opening with browser (cuz im a Wine user lol)
 - Difficulty Faces!